api/requirements.txt
asgiref==3.5.2
astroid==2.12.13
certifi==2022.9.24
cloudinary==1.30.0
dill==0.3.6
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
platformdirs==2.5.4
PyJWT==2.6.0
pylint==2.15.6
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.6
sentry-sdk==1.11.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
urllib3==1.26.12
wrapt==1.14.1
python-dotenv==0.21.0
drf_yasg==1.21.4